Add CSV Loadtest Recorder
